requirements.txt
awkward==2.9.0
h5py==3.15.1
loguru==0.7.3
mplhep==1.1.2
pandas==3.0.1
pennylane_lightning_gpu==0.44.0
pip==26.0.1
PyYAML==6.0.3
rdkit==2025.9.6
torch-geometric==2.7.0
torchvision==0.25.0
